Change verification logic for PGP/MIME messages
PGP/MIME can be signed in two places, embedded signature
in the pgp message, and MIME signature in the plaintext.

We now consider the message signature to fail only if both
of these signatures fails.
This means that if at least one of the signature verify correctly
we return the decrypted message with no errors.