Allow producing multiple singatures with a signing keyring
Up to this commit, GopenPGP selected a single signing key from the internal keyring,
i.e., calls (KeyRing).getSigningEntity().
Thus, it does not allow a user to create multiple signatures.
This commit changes the API such that each entity in the signing keyring given to a handle
produces a signature.