feat: Improve API for encryption with a session key
With RFC 9580, clients can encrypt data using AEAD with SEIPDv2 packets.
However, SEIPDv2 packets are only compatible with PKESK v6 and
SKESK v6 packets when the session key is encrypted using an OpenPGP certificate.

This commit improves the API to reduce the risk of encrypting data with SEIPDv2
while using a session key encrypted in a packet version below v6.
Specifically, the session key now includes an indicator of whether it is intended for AEAD use or not
and affects the produced packets.